Update global-alerts.css
Fixed an issue where the global-alerts were not working beyond the basic --info background with no working icons. Deminified the source material and applied it to this file. Added patch to enable bold or strong text to class .uq-alert--message: just use either <strong> or <b> to enable bolded text in an alert.